Separated some code out of event/models.py into event/processing.py and addressed comments
